Combine api and function evals
